Fix ~ usage - I'm not sure if it was wrong before (it did work), but at least now I know it's correct. :)
--HG--
branch : HEAD

+++ src/lib-index/mail-index-update.c
@@ -114,7 +114,7 @@ static size_t get_max_align_size(size_t base, size_t extra, size_t *max_extra)
 	extra = MEM_ALIGN(extra);
 	if (extra > *max_extra) {
 		/* partial */
-		extra = *max_extra & ~(MEM_ALIGN_SIZE-1);
+		extra = *max_extra & ~(size_t)(MEM_ALIGN_SIZE-1);
 		i_assert(extra <= *max_extra);
 	}
